requirements.txt
boto3==1.35.57
pandas==2.2.0
requests==2.32.3
pyarrow==18.0.0
apache-airflow==2.10.3
apache-airflow-providers-google==10.1.0